requirements.txt
fastapi
uvicorn
pydantic
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2